requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
conda==4.3.16
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.4
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-import-export==2.4.0
django-js-asset==1.2.2
django-smart-selects==1.5.9
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.5
pgi==0.0.11.2
Pillow==8.0.1
pipenv==2020.8.13

pycosat==0.6.3
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.25.10
vext==0.7.4
vext.gi==0.7.4
virtualenv==20.0.31
virtualenv-clone==0.5.4
Werkzeug==1.0.1
whichcraft==0.6.1
xlrd==1.2.0
xlwt==1.3.0
